pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vaultkeeper"
version = "0.1.0"
description = "Media drive cataloging and management system"
authors = [
    {name = "Dan Graetz", email = "dan@graetzmedia.com"},
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "pillow>=9.0.0",
    "qrcode>=7.0",
    "tqdm>=4.64.0",
    "pyyaml>=6.0",
    "python-magic>=0.4.27",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "isort>=5.0.0",
]
gui = [
    "pygobject>=3.40.0",
]

[project.scripts]
vaultkeeper = "vaultkeeper.cli.main:main"

[tool.setuptools]
packages = ["vaultkeeper"]
package-dir = {"" = "src"}

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true